+++ Package.swift
@@ -18,8 +18,8 @@ let package = Package(
     dependencies: [
         .Package(url: "https://github.com/Alamofire/Alamofire", majorVersion: 4),
         .Package(url: "https://github.com/ReactiveCocoa/ReactiveSwift", majorVersion: 1),
-        .Package(url: "https://github.com/ReactiveX/RxSwift", majorVersion: 3),
-        .Package(url: "https://github.com/antitypical/Result", majorVersion: 3)
+        .Package(url: "https://github.com/ReactiveX/RxSwift", majorVersion: 3, minor: 3),
+        .Package(url: "https://github.com/antitypical/Result", majorVersion: 3),
     ],
     exclude: [
         "Tests"