+++ Changelog.md
@@ -3,6 +3,8 @@
 - Fixed `reversedPrint` arguments for output.
 - **Breaking Change**: Added support to get the response (if any) from `MoyaError`.
 - Fixed memory leak when request with stub
+- Fixed memory leak when request with stub.
+- Changed `Moya.Error` to `MoyaError` in `MoyaAvailablity` for Swift 3.1 compatibility.
 
 # 8.0.2