+++ Changelog.md
@@ -1,5 +1,7 @@
 # Next
 
+- **Breaking Change** Updated `RxMoyaProvider.request` to return a [`Single<Request>`](https://github.com/ReactiveX/RxSwift/pull/1123)
+
 # 8.0.3
 
 - Fixed `reversedPrint` arguments for output.